Nerf the recharge rate of nameplate request slots
As this is meant to be a passive way of requesting profiles reducing
the recharge rate gives larger profile transfers a bit more breathing
room in high-population scenarios. With this change you'll now need to
wait 90 seconds for requests to be sent to 40 on-screen nameplates
instead of 60.

The wait period for on-screen nameplates before requests are sent has
also been increased to 3 seconds, in the spirit of Classic WoW mount
cast times.