Implement pooled queue strategy for profile comms
This reverts the previous change to comms in 2.3.3 which has, from
user feedback on Moon Guard, made things worse instead of better.

The new change implements a pair of queue strategies - independent and
pooled - which are selected based on the data being sent and the
approximate size of the data.

For all comms except sending profile data the independent strategy
is used, which leaves queuing down to the underlying transport library.

When sending profile data, if the size of the players' profile is above
a certain threshold of messages the pooled strategy will be enabled.
This uses a fixed number of queues for sending data, which should
reduce cross-addon contention (as well as for metadata comms) and
not allow individual profile transfers to choke one another out, but
for larger profiles this comes at the expense of increased latency
data is sent to any single player should the queues become large.