Implement smart nameplate queue strategy (#566)
* Implement smart nameplate queue strategy

This reworks the nameplate request system to use a much smarter queue
mechanism than we currently have.

Our existing approach only applies a cooldown on a per-character level,
such that we only throttle requests sent to nameplates of units we've
already seen at most once per 90 seconds. Unfortunately it doesn't
account for a few real-world issues.

Firstly, nameplates often quickly appear on-screen and then go off.
These can be thought of as one-hit wonders for people you're probably
not going to interact with and whose data isn't quite so important.

Secondly, you can walk into a field of a hundred players and turn your
nameplates on and suddenly request every visible nameplates' profile
immediately all at the same time. This completely throttles comms
on both sender and receiver ends.

To resolve this a smarter queue system is implemented for nameplate
requests with the following characteristics.

To solve the first issue, an enqueued request for a nameplate must first
exceed a minimum eligibility time before the request is sent. This
is kept small - at around 2.5 seconds - and should solve the issue of
nameplates briefly popping onto the screen an disappearing issuing a
full request for their data.

For the second issue, the queue system has an internal semaphore
representing request "slots". Each attempt to send a request will
first attempt to obtain a slot and decrement an internal counter; if
no slots are available the request won't be dispatched immediately and
must wait until a "recharge period" has elapsed. Each tick of the period
regains a single request slot.

The idea behind this is to significantly hamstring the ability to burst
request nameplate information; if you walk into a field of a hundred
players who all stay on screen to exceed their eligibility period, you
will only send at-most five requests to people on-screen immediately and
then one additional request each time a slot is regained every ~1.25s
thereafter.

Finally, the existing system of applying a per-character cooldown is
kept around - so if you repeatedly see a nameplate for someone you've
already requested then they will gain a significantly increased
eligibility time. The timer for this has been increased from 90 seconds
to 5 minutes; the idea is if you're interacting with someone for that
long you'll probably have seen their tooltip accidentally and requested
updated data anyway, or that they won't have made any change to their
profile worth querying.

The changes described significantly reduce the amount of requests that
can be sent in any short period of time, however for the user experience
this will come at a significantly increased latency as far as seeing an
undecorated nameplate and actually receiving their profile data.

To minimize this, there also exists a basic system of prioritized
requests. Prioritized requests will always be placed at the front of
the queue and apply to the nameplates of units that are either in your
friends list - both character and Battle.net, units that you are
presently grouped with, and members of your own guild.

* Minor changes to tweakable variable names

Primarily this just renames a few things around the cooldowns for
requests for people who haven't yet been submitted and for those
that have had requests submitted.

* Nerf the recharge rate of nameplate request slots

As this is meant to be a passive way of requesting profiles reducing
the recharge rate gives larger profile transfers a bit more breathing
room in high-population scenarios. With this change you'll now need to
wait 90 seconds for requests to be sent to 40 on-screen nameplates
instead of 60.

The wait period for on-screen nameplates before requests are sent has
also been increased to 3 seconds, in the spirit of Classic WoW mount
cast times.

* Dequeue pending requests on register data updates

If we receive updated data in our register for a unit, we'll dequeue
any pending request under the assumption that it was probably requested
by some other means.

* Add basic prioritization of nameplates by range

Keyword is "basic" - we just follow DBM's style of range check by
testing if a player is in range of a given item ID; in this case the
Mistletoe item is used because it's available on all game versions
and is relatively small at 23 yards.