Throttle information requests from "bad" versions
This should reduce the immediate need for everyone to update to 2.3.5
to gain from the comms fixes in this version; if we receive information
requests frequently from players running known bad version we'll not
respond to them unless three minutes have passed since the last
request.

This hack can probably be removed once 2.3.3/2.3.4 are out of widespread
use.