Disallow nameplate requests for distant units
This changeset removes the ability for nameplates to generate requests
for profiles beyond a 35 yard range, and applies a scaling cooldown
based on distance from the player to the unit so that closer units
(within ~10 yards) are requested sooner than those within 20, and those
ultimately within 35 yards.

With the values as configured, units within 10 yards will have an
initial request cooldown of 4s before the request is sent. This
increases at a rate of 0.4s per additional yard, up to 35 yards which
is 14s.

This changeset additionally refactors the request queue to split the
logic up now that everything had been implemented and all of the
requirements are known, and normalizes some verbiage in the nameplate
module elsewhere.