Avoid use of (removed) tooltip backdrop APIs
In patch 9.1.5 the tooltip templates no longer inherit support for
backdrops by default, and instead use the nineslice interface to
decorate themselves.

This change causes breakages in two areas:

	1. The chat link tooltip errors on load.
	2. Mousing over a player causes the character tooltip to error.

Blizzard has provided a partial deprecation wrapper which includes
basic support for the SetBackdrop method on tooltips still, but doesn't
extend this to cover the SetBackdropColor and SetBackdropBorderColor
methods.

To work around this we've introduced a new tooltip template for use
internally - TRP3_TooltipTemplate - which inherits a mixin providing
two new methods; SetBorderColor and SetCenterColor. These new method
names match new APIs present in 9.1.5 for nineslice panels.

Calling these new methods will delegate the application of the color
to an appropriate API for the current client automatically; on 9.1.5
it'll use the new methods on the nineslice panel, and on older clients
we'll use the existing backdrop methods which are still present on
the frame.

With the above in place, compatibility is fully restored and we only
need to rename a few function calls to use the new mixin methods.