Remove TRP3_BackdropTemplate
With the impending release of 1.14, we no longer need the polyfill
mixin and template for backdrop support in the Classic Era client. As
a result, we can remove this template and just use Blizzards' built-in
BackdropTemplate directly.

The custom backdrop tables have been kept around as they're specific
to our needs, and 9.1.5 complicates things by removing a bunch of the
ones we'd otherwise be using.

Note that this change would break Extended - it'll need to do the same
find/replace for TRP3_BackdropTemplate => BackdropTemplate.