Improve TextArea scroll behavior when editing
This modifies the TextAreaBase template to use Blizzards' ScrollingEdit
APIs which behave slightly better than our existing code with respect to
scrolling-down a scrollframe when the user is editing text toward the
lower end of the view.

The main changes this results in is that:

- Scrolling is no longer dependant upon the caret being at the end of
  the input field, and will occur if - for example - you're inserting
  text before it that would cause the new caret position to be outside
  the viewable field.

- Scrolling will leave a buffer region below the caret, so you've got
  a few visible lines for the mid-field edit case.

Any usages of the TRP3_TextAreaBase or TRP3_TextArea templates in
extended will need to be checked to see if they're providing an
OnLoad script handler, as one has been added to the template to
handle resizing of the child editbox component.