Correctly update text area layout on size changes
The "recent" changes to the TRP3_TextAreaBase template omitted the
script handler for OnSizeChanged; the method was defined on the mixin
but it wasn't actually hooked up properly.

This would lead to a few odd situations, including in Extended where
all the text areas had zero width and were invisible.

The OnSizeChanged handler has been added now, along with one for OnShow
that - while it shouldn't be needed - is being thrown in just as a
safeguard against any other untested scenarios.