Allow customizing Currently/OOC tooltip text color
This adds a new option to the tooltip settings that gives users a
defined way to customize the color of the currently and OOC text
contents in the tooltip, without needing to put in special characters
and breaking things for everyone else.

Only one option has been added and is shared between the two; I don't
see much purpose in having separate color pickers for currently and
OOC - especially when other settings are already linked between the
two, such as max line count.