Change color settings to title/main/secondary
This gives more flexibility over the whole tooltip without requiring
any per-field options.

The "title" color is used exclusively for the full title text on
player and companion tooltips.

The "main" color is used for most text in the tooltip, including
the race and level text as well as the section headers for currently
and OOC.

The "secondary" color is used for subtexts - such as the currently/OOC
text itself, or the name of the players' zone.