Fix extraneous linebreaks after block elements
Fixes #588 (probably). When converting things to HTML we end up
leaving a lot of stray newline characters at the start and/or end
of block-level elements, which causes inconsistencies as described
in the ticket.

This fix attempts to normalize newlines a little bit; if there is
a newline immediately proceeding a ">" character (before inline image
elements have been inserted) we assume this is a newline following
either the opening or closing tag pf block-level element such as a
paragraph or header. In this case, we chomp the (single!) newline
character and don't convert it to a break element.