Fix tooltip coloring with certain unicode inputs
The GameTooltip:AddLine() API appears to improperly apply the supplied
color to the displayed text if given an input string that contains
characters from specific unicode character sets, such as "𣆀".

We can work around this by wrapping each line in color sequence markup
to force the color to be applied. By doing so, we supply the color
white to the AddLine API as it's no longer needed.