Allow "/trp3 profile" macro conditionals
This commit adds the ability to use macro conditionals in the
"/trp3 profile" command, such as the following:

	/trp3 profile [combat] Profile A; Profile B

This only uses the WoW-builtin option parser so doesn't have the
advanced features we previously wanted for automation like checking
IC/OOC status, but this should enable _some_ use cases as we're
probably not going to implement that any time soon.

This may break existing usages of the command for profiles that
begin with an "[", but those should be few and far between you'd hope.