Fix guild membership line tooltip nil errors
When mousing over a player that's a member of a guild but has no
IC or OOC guild membership flags set in their profile the "membership"
text would be nil, cascading to an issue when generating the actual
tooltip contents.

Technically this bug has existed for a while as a nil membership text
*could* have prevented the fontstrings being created, leading to an
issue with assigning fonts - it just wasn't too likely as you'd need
to have mouseovered them as the first player in the session.