Fix spurious crash when dragging Misc. Info fields
If the asset supplied to SetCursor exceeds 64x64 then the client will
fail to display the cursor, and instead uses the previous one that was
assigned. If there is no previous cursor - because for example the user
dragged a field right after starting the game - it'll crash with a
segfault from reading a near-null address.

Our workaround is that if the asset exceeds 64x64 we'll instead show
a question mark icon while dragging to avoid crashing. Note that due
to _another_ bug, we need to use the file ID for the question mark
asset since for some inexplicable reason a file path causes it to
display at completely the wrong size.