Manage external libraries through the packager
The submodule approach for library management isn't ideal and has
issues with maintenance. We're shipping old versions of libraries
without updating them periodically, and in many cases are having
to hard-embed them due to the lack of an external git repository
from which to source them.

Addon packaging approaches use the .pkgmeta to handle external
libraries as a first-choice, falling back on hard embedding only
when otherwise absolutely impossible.

As such the packaging has been reconfigured to remove all of our
submodule-managed libraries, leaving only our copy of Ellyb,
MSA-DropDownMenu, TaintLess, and UTF8 as these libraries are
either self-managed or have no external repository.

The advantage of all this is we don't need to do the submodule
dance now every time we update Chomp, LibMSP, or LibRPMedia.

One downside is that - for development - none of the required libraries
will be present by default. There'll be a script added soon to fix
this issue, but other projects are "just fine" with that situation
since all of the libraries in question can just be installed as
standalone addons for development purposes.