Remove Taintless library
This doesn't work as-is in Dragonflight and we don't technically have
permissions to modify it. The fixes themselves aren't that major now
anyway, as retail should be relatively insulated from dropdown related
taint. Classic is another story, but given how the guild controls
stuff is broken there anyway even with Taintless installed it seems
unlikely that this library was really accomplishing its goal.