Loosen requirements on pet unit checks
On Classic Wrath the Water Elemental pet summoned by mages lacks
a valid return for UnitCreatureFamily, causing getTargetType to fail
to identify Water Elementals as valid units for companion profile
eligibility.

As a workaround we'll loosen the restrictions on pet unit candidacy,
only requiring that the API identify a given unit as either being
the players' current pet unit or someone elses' pet. This _may_ have
unintended consequences, but it feels better than hardcoding an
exception for specific NPC IDs as these can differ based on glyphs
and such.

Fixes #634.