Implement more generic '/trp3 set' command
This enhances the original command beyond the scope of just modifying
"currently" text and now supports changing a subset of other profile
fields, as well as macro conditionals to control what values actually
get written out.

The following fields are supported:
  - class <text...>
  - classcolor #rrggbb
  - currently <text...>
  - firstname <text...>
  - fulltitle <text...>
  - icon <icon name>
  - lastname <text...>
  - oocinfo <text...>
  - race <text...>
  - title <text...>

For macro conditionals, the conditions come after the field name and
will select the value. These are only processed if the first character
of the value is an "[" so as to avoid breaking strings on semicolons.

Help text has been added for the '/trp3 set' command, which will also
print system messages if a requested field is unknown or is given
an invalid value.