Add profile name conditional
Profile names are compared in a normalized manner with all whitespace,
control characters, and punctuation converted to hyphens. Case folding
to lowercase is also performed to the extent possible.

When comparison takes place, we accept any partial match against the
current profile name - this is just because profile names by default
are quite long what with having a realm name prefix.