Implement automated CF localization import (#639)
* Implement automated CF localization import

This moves the handling of localization out of the packagers' hands
and integrates an automated GitHub action that'll obtain updated
translations, from which a PR will be created to merge in the latest
changes.

The advantages here are that we can more easily verify whether or not
any locales have broken format strings, as has happened a couple of
times now.

The action will trigger daily at 18:00 and create a PR if there's any
outstanding localization updates. If a PR already exists, it'll be
updated with the new contents - we won't get spammed with a million
separate PRs. Additionally the action will trigger on any commit to the
main branch (such as when merging in PRs) and can be triggered manually.

* Skip ec-check on all locale files

* Add initial locale files

* Make luacheck ignore generated locales

* Travis, I've come to bargain