Add "/trp3 set" command (#640)
* Add '/trp3 currently <text>' command

* Trim text prior to applying it

* Disallow currently changes for default profiles

* Improve chatframe feedback on currently changes

* Implement more generic '/trp3 set' command

This enhances the original command beyond the scope of just modifying
"currently" text and now supports changing a subset of other profile
fields, as well as macro conditionals to control what values actually
get written out.

The following fields are supported:
  - class <text...>
  - classcolor #rrggbb
  - currently <text...>
  - firstname <text...>
  - fulltitle <text...>
  - icon <icon name>
  - lastname <text...>
  - oocinfo <text...>
  - race <text...>
  - title <text...>

For macro conditionals, the conditions come after the field name and
will select the value. These are only processed if the first character
of the value is an "[" so as to avoid breaking strings on semicolons.

Help text has been added for the '/trp3 set' command, which will also
print system messages if a requested field is unknown or is given
an invalid value.

* Avoid hardcoding field names in localization

* Use 'Custom' naming scheme for class/race data

* Improve help text colorization

Additionally moved the example commands and usage text out of the
localized string data, so we can change the syntax later without
needing to re-localize things.

* Re-allow changing OOC status for default profile

Field modification for the default profile will be permitted for the
roleplay status field even if the current profile is a default one,
as this is how it has been for some time now.

* 'Same for RP experience'

* Add custom macro conditional evaluator

This implements a basic framework for custom macro conditionals and
provides an "ooc", "ic", and "rpstatus:<state>" trio of conditionals
that can be evaluated in option strings.

This has been hooked into the "/trp3 set" command so you can
conditionally set fields based upon IC/OOC state, and has also
been extended to apply to the previous work on "/trp3 profile".

* Remove conditional initialization hooks

These actually have issues if called after loading has completed. They
probably won't be needed, and fixing the issue adds complexity so we'll
omit them for now and reconsider later.

* Add profile name conditional

Profile names are compared in a normalized manner with all whitespace,
control characters, and punctuation converted to hyphens. Case folding
to lowercase is also performed to the extent possible.

When comparison takes place, we accept any partial match against the
current profile name - this is just because profile names by default
are quite long what with having a realm name prefix.

* Add location name macro conditional

This is available via both "[loc:<name>]" and "[location:<name>]". Like
the original PR this uses the minimap zone text, and as with profile
names is a normalized loose comparison that's case-folded and only
needs to be a partial match.

* Add macro conditional support to "/trp3 status"

As the profile and set command support it, it feels useful to also
extend this to status so you can have it set properly based on
stuff like location.

* Refactor set command implementation

Relocate to the automation module as it seems like a good fit. Also
improved handling of the help text generation and its localizability.

* Correctly generate hex colors