Dragonflight: Housekeeping (#631)
* Manage external libraries through the packager

The submodule approach for library management isn't ideal and has
issues with maintenance. We're shipping old versions of libraries
without updating them periodically, and in many cases are having
to hard-embed them due to the lack of an external git repository
from which to source them.

Addon packaging approaches use the .pkgmeta to handle external
libraries as a first-choice, falling back on hard embedding only
when otherwise absolutely impossible.

As such the packaging has been reconfigured to remove all of our
submodule-managed libraries, leaving only our copy of Ellyb,
MSA-DropDownMenu, TaintLess, and UTF8 as these libraries are
either self-managed or have no external repository.

The advantage of all this is we don't need to do the submodule
dance now every time we update Chomp, LibMSP, or LibRPMedia.

One downside is that - for development - none of the required libraries
will be present by default. There'll be a script added soon to fix
this issue, but other projects are "just fine" with that situation
since all of the libraries in question can just be installed as
standalone addons for development purposes.

* Add basic Makefile

The makefile can be used to pull in external libraries through the
packager, or to do an ad-hoc packaged build for testing.

For Windows platforms you'll probably want to run it through WSL;
eg. "wsl make libs" or "wsl make dist".

* De-clutter .editorconfig

* Remove appendix from LICENSE

We don't really need this to tell us how to apply the license when
just having the document itself applies the license.

* Normalize license headings

Moved copyright and author attributions to a new file at the root of
the repository. We now use the short variant of the Apache license
header in all files consistently, namely:

	Copyright The Total RP 3 Authors
	SPDX-License-Identifier: Apache-2.0

* Add XML schema validation to "make check"

We _may_ want to integrate this into CI checks too at some point, but
for now I'd recommend avoiding it as the XSD does have some potential
issues that need resolving.

* Fix XML validation errors

SimpleHTML apparently doesn't support Color child elements; unsure
if that's just a bad copy/paste or some historical leftover.

For the color picker layers this feels more like the XSD being too
strict, however the anchors were incorrectly defined.

* Remove unused workarounds script

While loaded we don't actually have any workarounds to apply. It's
also not really our place to be implementing workarounds for Blizzard
issues _generally_; would rather leave that to some other more
interested third-party.

* Normalize filename casing

We've had an awkward mix of snake_case and PascalCase filenames for
a while now; all the newer case has transitioned to the latter but
the former was never really cleaned up.

This fixes that and renames everything to follow the PascalCase
scheme going forward.

* Make directory structure more flat

This primarily rearranges the "Core" folder to move its subdirectories
to the top-level of the addon. Basically everything outside of the
"Modules" folder is already assumed to be core - so we don't really
need to have such a massively nested directory structure within
"Core" making things confusing.

Of note:

  - The models folder has been integrated into Core as there's only
    the Player model for now.
  - The tests folder is now at the top of the addon directory.
    - Tests are only loaded in unpackaged builds now also.
  - The UI folder is now at the top of the addon directory.

* Remove LibSerialize

This was accidentally added based off a search since we've called
AceSerialize "libSerializer" somewhere in the code.

* Fix exclusion paths in luacheck config

This didn't flag for local testing as I'd been using "make check"
which was excluding the files through an ls-files filter removing
the properly-cased Libs directory.

* Fix ec-checker step in github actions

This also needed case correction for its own excludes. This should
be moved to "make check" really.

* Add some library optdepends

This is only for a few important libraries to ensure that updates
are picked up before we've loaded our addon. It isn't required as
these will upgrade automatically, but it is a bit more efficient.

* Add '-S' flag to 'make dist'

* Force LibMSP to use main branch temporarily

* Add all libraries as opt-depends