Dragonflight Patches (#633)
* Fix SetFont errors due to missing flags

* Fix SetText failing with nil HTML parameter

* Another nil HTML error

* Remove Taintless library

This doesn't work as-is in Dragonflight and we don't technically have
permissions to modify it. The fixes themselves aren't that major now
anyway, as retail should be relatively insulated from dropdown related
taint. Classic is another story, but given how the guild controls
stuff is broken there anyway even with Taintless installed it seems
unlikely that this library was really accomplishing its goal.

* Remove unnecessary (?) SetJustifyH calls

These error due to SimpleHTML widgets needing an element type (like
"p" or "h1") as the first parameter; as such these have probably
never worked and even if fixed appear to do nothing?

* Fix potential GetPoint errors with target frame

* Fix script error when opening URL popups

* Re-add TaintLess library

This has since been updated with fixes. Shame we can't package it
normally because someone needs to be a special bean but...

* Version cleanup (#637)

* Version cleanup

* Meodge

* Fix map scan animation crash

The Scale animations were configured to reference non-existent target
keys on the animation instance directly, rather than looking them up
in the parent hierarchy.

* Removing the scale

* Disable UnitPopup integrations for 10.0.0 (#638)

In an attempt to try and fend off any questions on why, the settings
page for the module will have some text added to explain the reasoning.

For development purposes (and for users who really don't care about
edit mode and will have fully custom UIs) a temporary config variable
has been added to bypass disabling the module.

* Remove branch name override for LibMSP

* Correct exclusion paths in pkgmeta

* Re-enable UnitPopups module in Classic

If the EditModeManagerFrame doesn't exist we'll assume it's not 10.0.0
and allow unitpopup integrations to work as normal. This adds a bit
more of a hack atop the existing stuff but hopefully it'll all be
resolved soon.

* Implement automated CF localization import (#639)

* Implement automated CF localization import

This moves the handling of localization out of the packagers' hands
and integrates an automated GitHub action that'll obtain updated
translations, from which a PR will be created to merge in the latest
changes.

The advantages here are that we can more easily verify whether or not
any locales have broken format strings, as has happened a couple of
times now.

The action will trigger daily at 18:00 and create a PR if there's any
outstanding localization updates. If a PR already exists, it'll be
updated with the new contents - we won't get spammed with a million
separate PRs. Additionally the action will trigger on any commit to the
main branch (such as when merging in PRs) and can be triggered manually.

* Skip ec-check on all locale files

* Add initial locale files

* Make luacheck ignore generated locales

* Travis, I've come to bargain

* Add "/trp3 set" command (#640)

* Add '/trp3 currently <text>' command

* Trim text prior to applying it

* Disallow currently changes for default profiles

* Improve chatframe feedback on currently changes

* Implement more generic '/trp3 set' command

This enhances the original command beyond the scope of just modifying
"currently" text and now supports changing a subset of other profile
fields, as well as macro conditionals to control what values actually
get written out.

The following fields are supported:
  - class <text...>
  - classcolor #rrggbb
  - currently <text...>
  - firstname <text...>
  - fulltitle <text...>
  - icon <icon name>
  - lastname <text...>
  - oocinfo <text...>
  - race <text...>
  - title <text...>

For macro conditionals, the conditions come after the field name and
will select the value. These are only processed if the first character
of the value is an "[" so as to avoid breaking strings on semicolons.

Help text has been added for the '/trp3 set' command, which will also
print system messages if a requested field is unknown or is given
an invalid value.

* Avoid hardcoding field names in localization

* Use 'Custom' naming scheme for class/race data

* Improve help text colorization

Additionally moved the example commands and usage text out of the
localized string data, so we can change the syntax later without
needing to re-localize things.

* Re-allow changing OOC status for default profile

Field modification for the default profile will be permitted for the
roleplay status field even if the current profile is a default one,
as this is how it has been for some time now.

* 'Same for RP experience'

* Add custom macro conditional evaluator

This implements a basic framework for custom macro conditionals and
provides an "ooc", "ic", and "rpstatus:<state>" trio of conditionals
that can be evaluated in option strings.

This has been hooked into the "/trp3 set" command so you can
conditionally set fields based upon IC/OOC state, and has also
been extended to apply to the previous work on "/trp3 profile".

* Remove conditional initialization hooks

These actually have issues if called after loading has completed. They
probably won't be needed, and fixing the issue adds complexity so we'll
omit them for now and reconsider later.

* Add profile name conditional

Profile names are compared in a normalized manner with all whitespace,
control characters, and punctuation converted to hyphens. Case folding
to lowercase is also performed to the extent possible.

When comparison takes place, we accept any partial match against the
current profile name - this is just because profile names by default
are quite long what with having a realm name prefix.

* Add location name macro conditional

This is available via both "[loc:<name>]" and "[location:<name>]". Like
the original PR this uses the minimap zone text, and as with profile
names is a normalized loose comparison that's case-folded and only
needs to be a partial match.

* Add macro conditional support to "/trp3 status"

As the profile and set command support it, it feels useful to also
extend this to status so you can have it set properly based on
stuff like location.

* Refactor set command implementation

Relocate to the automation module as it seems like a good fit. Also
improved handling of the help text generation and its localizability.

* Correctly generate hex colors

* Add "/trp3 location" command (#641)

This command allows toggling the state of player location broadcasts
from the chat frame, and integrates with the custom macro conditional
work that's just been done for "/trp3 set".

Co-authored-by: Solanya <17127080+Solanya@users.noreply.github.com>
Co-authored-by: Solanya <solanya@totalrp3.info>