Use C_TooltipInfo APIs for companion owner names
A future build of 10.0.2 will add a new UnitOwner tooltip line type that
provides the GUID of the owning unit as part of its payload. We can
use this to extract owner names and realm, thus removing the need
to de-format localization strings and deal with locale-specific
quirks like declensions.