Use taint-free hyperlink prefixes
The approach other addons use for local links is to take advantage of
"garrmission" link types with an addon name in the data segment after
the first colon. This is handled by SetItemRef as an invalid mission
link which then returns, and as such can be caught with a secure
hook without needing to insecurely modify or replace the
SetHyperlink method on ItemRefTooltip.