Cache configured locale in Data addon
This changeset modifies our handling of localization during early load.

Due to providing a configurable addon locale setting, we've got issues
where any localization lookups that occur before the ADDON_LOADED event
fire will always default to the client locale, which may differ from
the configured setting. This is particularly noticeable with module
registrations, but it also complicates the use of strings within UI
elements that get created early on.

While personally I'd quite like to remove the setting entirely due to
the massive complications this causes, we have a workaround at our
fingertips that resolves part of the problem.

The Data addon has gained a new saved variable (TRP3_AddonLocale) which
will mirror the AddonLocale setting when saved variables are written.
This is only set if the configured locale is not the client default,
otherwise it will be reset to nil.

On the following startup if the value stored in TRP3_AddonLocale is
not nil it will be honored by the localization system when early setup
occurs in the script body. If it is nil, it'll be set to the default
locale of the client and match todays' behavior for early startup.

As an extension to this, we also now honor the custom GAME_LOCALE
global variable. This is accepted by other localization addons for
the purposes of testing translations, but may also be set by addons
or scripts that intentionally mass-override locales for other addons.
The value assigned to this global has a higher priority than the
default client locale.

If we were to remove the setting, I suspect we'd really just need to
support GAME_LOCALE and call it a day.