Date Format Tweaks (#675)
* Add instructions field to config editboxes

This can be used for default values or placeholder text.

* Allow advanced date format configuration

Instead of using a dropdown we'll allow people to just outright
fully customize the format manually. If left as an empty string,
a default based upon the configured addon locale will be used
appropriate to external references for worldwide date formats.

* Use GetPreferredLocale function elsewhere

Also remove a few bits of code that were listening to locale changes;
this isn't ever going to take off :)

* Clean up unused vars

* Lowercase 'Format' in setting name

* Add safety wrapper for generating date strings

As some date format inputs are invalid ("%N"), a new utility to
format a timestamp/time table to a date string has been added which
will fall back to the default locale-specific date format if an error
occurs during formatting.

* Clarify help text for other characters

* Use securecall for protected calls

This allows the debug locals dump to include the problematic
format string.