Pull in UTF8 library through packager externals
This library has a source repo so doesn't require hard embedding;
the only reason it wasn't included in the initial library migration
work was because of our custom lib.xml file.

That lib.xml file is now committed upstream so we can just import the
library the "normal" way and have one less hard-embed floating about.