Plater Nameplates module necromancy
Plater now allows developers to import mods from within their own addons, so that is the goal of this module. It currently works and allows hot reloading of the module without needing to reload the UI.

The whole 'hot reload' system is not great and probably is not well made, but it works for now. Other modules can support it by setting 'hotReload' to `true` in the module.