Remove the English Language
With the autoimport of localization data we've got a bit of a weird
setup whereby our default localization texts are English but then get
replaced by the contents of the English "translations" imported from
CurseForge.

That system leaves a bit of room for localization strings to get
out-of-sync for English, so this changeset closes that hole by:

  1. Removing import of the "enUS" locale from CF.
  2. Moving the default localization table to the "enUS" locale script.

Additionally the processing of pet owner matching patterns has been
relocated to the UITools script and made a less... awkward.

  1. Matching strings no longer care about the game locale; the same
     set of transformations is applied across all locales.
  2. For French locales, "|2 " is now replaced with a single pattern
     for matching both "d'name" and "de name".
  3. For French locales, the hack for not accidentally matching
     companion pet type/level lines has been removed. Instead, the
	 patterns just have start/end anchors which won't match those lines.
  4. When matching lines we now scan all the tooltip lines rather than
     hardcoding guesses that then differ based upon client flavor or
	 colorblind mode.
  5. The squire title type has been added to support the Argent Squire
     companion pet - just in case people want to call him Billy
	 or something.