Add workflow for automated CF localization uploads
At-present this workflow is configured for manual execution; it has
been tested successfully in a dummy project and should be safe for
general usage.

Note that the "delete missing phrases" functionality isn't all that
dangerous; CF doesn't delete phrases but simply marks them as "deleted"
which prevents them showing up in the web UI by default and prevents
them being exported. Un-deleting phrases is possible and won't lead to
a loss of existing translations.