Persist name-only mode state in local settings
An upcoming client build will be restoring the name-only mode cvar
however won't persist it across client restarts, meaning that we'll
need to store its state ourselves and restore it on load.

This commit attempts to implement that in such a way that we'll
hopefully avoid confusing users and won't conflict with any other
addons attempting to control the cvar.

The checkbox in the UI will represent the current effective state
of name-only mode, showing a check if the cvar is enabled or not. If
the state of the cvar doesn't match our locally saved state, the
check will be desaturated - similar to how tristate checkbuttons work
in the addon list. Clicking the checkbutton operates on our locally
saved state, toggling based off that and not the effective one.

The cvar itself will only ever be enabled from us at runtime; if the
user attempts to disable name-only mode within our settings they'll
need to do a UI reload for the change to take effect. This is the same
as the existing behaviour anyway, however the specific reason we won't
toggle the cvar back to an off state is to avoid edge cases where we
disable name-only mode after other addons in the environment have also
tried to enable it.

On addon load we'll only set the cvar for name-only mode from within
the Blizzard nameplates decorator. As such - if the user has Plater
or Kui installed, we'll leave it to them to configure name-only mode.
This might change based upon feedback.