Only show reload prompt when disabling
At some point in Mainline (probably with the 10.0 settings changes?)
the name-only mode cvar doesn't actually require a UI reload to take
effect.

As however we don't explicitly ever unset the cvar to avoid addon
conflicts, we do still need to prompt for a reload when the user
disables the settings in our options panel.