Plater Nameplates Support (#657)
* Adding primitive support for Plater Nameplates integration

Proof of concept, not usable at this time. Does however, show that this is possible.

* Fixed non-RP unit nameplates

Non-default plater profiles do not work with this module. I'll sleep on it and use the nightmares to solve the problem tomorrow.

* Guild name support, proper name coloring

This is mostly playable now, as long as you don't need all the options to function, and are using the default Plater profile.

* Removed increased contrast check

Removed temporarily. It has a tendency to crash the game in populated areas.

* Adding icon support, full title strings

This version is much more usable than the last two - still requires you to use the default Plater profile, but everything* should be functional.

If anyone is reading this, I'd love feedback on the appearance of the nameplates!

* Currently the "use class color" setting is not working.

* Addressing feedback from Meo

* Fixing inconsistent hex color in Locale and an "in-combat" error

* Yeeting crimes and high-contrast adjustments

We don't need to manually account for high-contrast as it's handled automatically by the color provided by Nameplates_Core

* Updating locale

When the nameplate settings are updated, the nameplates are refreshed, but the visibility settings don't apply properly until nameplates are regenerated.

Will revert as needed.

* Added tentative support for non-default Plater profiles

I also made some of the settings a little bit more responsive by properly refreshing the frame when the configuration settings are updated.

* Yeeting integrated Plater module

I'm keeping the file in place just in case, but most of the functionality has been moved to a standalone Plater mod.

Also keeping localization changes because they still need to exist.

* Cleansing the sins of the past

Removing this file as it is no longer needed with the functionality being moved to a standalone Plater mod.

* Removing module localization

This is no longer needed as Plater Nameplates will no longer be a TRP module.

* Plater Nameplates module necromancy

Plater now allows developers to import mods from within their own addons, so that is the goal of this module. It currently works and allows hot reloading of the module without needing to reload the UI.

The whole 'hot reload' system is not great and probably is not well made, but it works for now. Other modules can support it by setting 'hotReload' to `true` in the module.

* Updating import string (for now)

* Scaring off luacheck

hopefully

* Yeeted spaces into space.

* Updated .luacheckrc and removed erroneous print()

This should function now, featuring module hot-reloading!

* Fixed wording and updated import string

* Fixed mod importer flooding the garbage can

* Added some (hopefully) helpful comments

* Allow modules to specify external addon dependencies

By using the string "external" in place of the dependency version, you can specify an external addon as a dependency.

* Fixed mild oversight

* Changed minVersion to be less confusing

What does this even do?

* Definitely not another oversight

* Fixed missing module dependency

* Cleaned up hot reload functionality

Modules will also show on their tooltips if they support hot reload.

* Removed imposter

* Spring cleaning

* Fixed Plater mod not being re-enabled on reload

* Fixed module onStart() being called erroneously

* Updated Plater mod to update nameplates as new profile data is received

I can't believe it's not butter! I mean... I can't believe I didn't think of this sooner.

* Tick nameplates in onStart() to update them immediately with RP names

* Removed mod prio setting

Changing this externally increments the version of the mod, causing it to send itself to the 'recycle bin' when it's reimported later, eventually filling the entire recycle bin with TRP3 Nameplates, and we all know that's not recyclable.

* Professionalism

* Helpful notes for the future

* Cleanup

* Addressing the controversy

* for my next magic trick I'll make those lines reappear

---------

Co-authored-by: Solanya <solanya@totalrp3.info>