Hide OOC characters on map scan by default (#696)
Closes #475. This adds a new configurable option to hide OOC players
from the map scan results. Per the ticket, this option is enabled
by default.

For compatibility with existing clients on the network we treat all
incoming scan responses that don't supply a roleplay status value
as being in-character.

To additionally differentiate between IC and OOC characters in the
scan results, the pin tooltip will now additionally display an "(OOC)"
suffix for characters. For lone pins, they will also be faded out.

Co-authored-by: Solanya <solanya@totalrp3.info>