Add ROLEPLAY_STATUS_CHANGED event
This acts as a filter on the existing REGISTER_DATA_UPDATED event
rather than being something that is manually triggered.

Users of this event should not rely on any sort of ordering between
callbacks associated with either this event or REGISTER_DATA_UPDATED;
it is perfectly valid for the ordering of such callbacks to be that
either of them can be invoked before the other.