Refresh toolbar buttons more aggressively
We'll update the toolbar buttons with a ticker rather than an OnUpdate
script and also trigger an update on status changes. Won't commit to
anything more flexible than this for now.

Additionally when refreshing toolbar buttons we'll now invalidate the
time-since-last-update on the individual buttons, which should trigger
their "onUpdate" logic at the end of the current frame.