Use ROLEPLAY_STATUS_CHANGED for toolbar visibility
This additionally loosens the effect of the manual toolbar toggle such
that it only affects the immediate visibility of the bar.

If the user has the "Only show while in character" setting selected and
is out of character, toggling the toolbar will show it. If they now
change status twice (IC then OOC) it'll hide rather than being stuck
shown until a UI reload or config change.