Additional toolbar visibility options (#699)
* Adding additional toolbar visibility options

Instead of just choosing whether or not to show the toolbar on login, users can now choose between showing on login, showing only while in-character, or they can hide the toolbar altogether.

* Optionamus Responsus

The `toolbarVisibility` local variable wasn't being updated within the event listener, so it wouldn't reflect config changes without a reload.

* Toolbar cleanup

* String Theory

* Lamb Sauce

* Yeeting help string to the great string table in the sky

The options should be self-explanatory. Can re-add if it's something we need later down the line.

* The Toolbar Show

* Extra Lamb Sauce

lamb sauce = semicolons

* Heresy

Adding functionality to lock out the automatic hiding/unhiding of the toolbar when you use the slash command or the minimap button to toggle the toolbar visibility.

* String Necromancy

The help string returns newly invigorated and ready to assist those in need of knowledge.

* Rework state management into a new Toolbar mixin

Now that we've got an idea on how we want this setting to actually
work it makes most sense to just move all the state management to
a mixin rather than spattering it throughout half the source file.

* Update help text to my suggestion

Clearly my suggestion is the best since no one's countered it for at
least six hours now, so I'm proclaiming myself to be king here.

* Add ROLEPLAY_STATUS_CHANGED event

This acts as a filter on the existing REGISTER_DATA_UPDATED event
rather than being something that is manually triggered.

Users of this event should not rely on any sort of ordering between
callbacks associated with either this event or REGISTER_DATA_UPDATED;
it is perfectly valid for the ordering of such callbacks to be that
either of them can be invoked before the other.

* Small documentation cleanup

* Use ROLEPLAY_STATUS_CHANGED for toolbar visibility

This additionally loosens the effect of the manual toolbar toggle such
that it only affects the immediate visibility of the bar.

If the user has the "Only show while in character" setting selected and
is out of character, toggling the toolbar will show it. If they now
change status twice (IC then OOC) it'll hide rather than being stuck
shown until a UI reload or config change.

* Refresh toolbar buttons more aggressively

We'll update the toolbar buttons with a ticker rather than an OnUpdate
script and also trigger an update on status changes. Won't commit to
anything more flexible than this for now.

Additionally when refreshing toolbar buttons we'll now invalidate the
time-since-last-update on the individual buttons, which should trigger
their "onUpdate" logic at the end of the current frame.

* Misc. status read cleanup

* Update target bar visibility on status changes

Per Discord this functionality was behaving counterintuitively; if the
user had a target selected then toggling roleplay status wouldn't
show or hide the target bar.

* Reuse toolbar visibility strings for target frame

The settings are effectively identical anyway. Personally I prefer the
newer strings over the old so...

* Small tweaks to localization strings

The lockout string is no longer used. The "Toolbar Visibility" option
is now "Display conditions" to match with the existing text for the
same options exposed by the target frame.

* Removed the lockout portion of the help text

* Temp flyway needs proper enum access

* One more changelog line for the road

* Globalization (of toolbar visibility options)

* Flyway version update

* Remove useless line

---------

Co-authored-by: Daniel Yates <me@meorawr.io>
Co-authored-by: Solanya <solanya@totalrp3.info>